+++ packages/yourloops/components/device/cgm-info-table.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Diabeloop
+ * Copyright (c) 2023-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,159 +25,51 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { type FC } from 'react'
 import type { CgmConfig } from 'medical-domain'
 import { CGMName } from 'medical-domain'
-import ListItem from '@mui/material/ListItem'
-import ListItemText from '@mui/material/ListItemText'
-import Typography from '@mui/material/Typography'
+import React, { type FC } from 'react'
 import { useTranslation } from 'react-i18next'
-import Box from '@mui/material/Box'
-import { GenericListCard } from './generic-list-card'
 import { formatDateWithMomentLongFormat } from '../../lib/utils'
+import { GenericListCard } from './generic-list-card'
 
 interface CgmInfoProps {
   cgm: CgmConfig
 }
 
-const FALLBACK_VALUE = '-'
-
 export const CgmInfoTable: FC<CgmInfoProps> = ({ cgm }) => {
   const { t } = useTranslation()
 
+  const getTableLines = (cgm: CgmConfig): { value: string, label: string }[] => {
+    const tableInfoLines = [
+      { label: t('Manufacturer'), value: cgm.manufacturer },
+      { label: t('Product'), value: cgm.name },
+      { label: t('Cgm sensor expiration date'), value: formatDateWithMomentLongFormat(new Date(cgm.expirationDate)) }
+    ]
+
+    const g6InfoLines = [
+      { label: t('Cgm transmitter software version'), value: cgm.swVersionTransmitter },
+      { label: t('Cgm transmitter id'), value: cgm.transmitterId },
+      { label: t('Cgm transmitter end of life'), value: formatDateWithMomentLongFormat(new Date(cgm.endOfLifeTransmitterDate)) }
+    ]
+
+    const g7InfoLines = [
+      { label: t('serial-number'), value: cgm.serialNumber },
+      { label: t('software-number'), value: cgm.softwareNumber },
+      { label: t('pairing-code'), value: cgm.pairingCode }
+    ]
+
+    const additionalLines = cgm.name === CGMName.G6 ? g6InfoLines : g7InfoLines
+    tableInfoLines.push(...additionalLines)
+
+    return tableInfoLines
+  }
+
   return (
-    <GenericListCard title={t('CGM')} data-testid="settings-table-cgm">
-      <ListItem divider className="list-item">
-        <ListItemText>
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between"
-            }}>
-            <Typography variant="body2">{t('Manufacturer')}</Typography>
-            <Typography variant="body2" className="bold">{cgm.manufacturer ?? FALLBACK_VALUE}</Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-      <ListItem divider className="list-item">
-        <ListItemText>
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between"
-            }}>
-            <Typography variant="body2">{t('Product')}</Typography>
-            <Typography variant="body2" className="bold">{cgm.name ?? FALLBACK_VALUE}</Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-      <ListItem divider className="list-item">
-        <ListItemText>
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between"
-            }}>
-            <Typography variant="body2">{t('Cgm sensor expiration date')}</Typography>
-            <Typography variant="body2" className="bold">{formatDateWithMomentLongFormat(new Date(cgm.expirationDate)) ?? FALLBACK_VALUE}</Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-      { cgm.name === CGMName.G6 &&
-        <>
-        <ListItem divider className="list-item">
-        <ListItemText>
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between"
-            }}>
-            <Typography variant="body2">{t('Cgm transmitter software version')}</Typography>
-            <Typography variant="body2" className="bold">{cgm.swVersionTransmitter ?? FALLBACK_VALUE}</Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-      <ListItem divider className="list-item">
-        <ListItemText>
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between"
-            }}>
-            <Typography variant="body2">{t('Cgm transmitter id')}</Typography>
-            <Typography variant="body2" className="bold">{cgm.transmitterId ?? FALLBACK_VALUE}</Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-      <ListItem>
-        <ListItemText className="list-item">
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between"
-            }}>
-            <Typography variant="body2">{t('Cgm transmitter end of life')}</Typography>
-            <Typography
-              variant="body2"
-              className="bold"
-            >
-              {formatDateWithMomentLongFormat(new Date(cgm.endOfLifeTransmitterDate)) ?? FALLBACK_VALUE}
-            </Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-        </>
-      }
-      { cgm.name !== CGMName.G6 &&
-        <>
-          <ListItem divider className="list-item">
-            <ListItemText>
-              <Box
-                sx={{
-                  display: "flex",
-                  justifyContent: "space-between"
-                }}>
-                <Typography variant="body2">{t('serial-number')}</Typography>
-                <Typography variant="body2" className="bold">{cgm.serialNumber ?? FALLBACK_VALUE}</Typography>
-              </Box>
-            </ListItemText>
-          </ListItem>
-          <ListItem divider className="list-item">
-            <ListItemText>
-              <Box
-                sx={{
-                  display: "flex",
-                  justifyContent: "space-between"
-                }}>
-                <Typography variant="body2">{t('software-number')}</Typography>
-                <Typography
-                  variant="body2"
-                  className="bold"
-                >
-                  {cgm.softwareNumber ?? FALLBACK_VALUE}
-                </Typography>
-              </Box>
-            </ListItemText>
-          </ListItem>
-          <ListItem>
-          <ListItemText className="list-item">
-            <Box
-              sx={{
-                display: "flex",
-                justifyContent: "space-between"
-              }}>
-              <Typography variant="body2">{t('firmware-version')}</Typography>
-              <Typography
-                variant="body2"
-                className="bold"
-              >
-                {cgm.firmwareVersion ?? FALLBACK_VALUE}
-              </Typography>
-            </Box>
-          </ListItemText>
-        </ListItem> 
-        </>
-      }
-    </GenericListCard>
+    <GenericListCard
+      title={t('CGM')}
+      tableLines={getTableLines(cgm)}
+      data-testid="settings-table-cgm"
+    />
   )
 }
+