+++ packages/yourloops/components/device/generic-list-card.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2025, Diabeloop
+ * Copyright (c) 2023-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,17 +25,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { type FC, type PropsWithChildren } from 'react'
 import Card from '@mui/material/Card'
-import CardHeader from '@mui/material/CardHeader'
 import CardContent from '@mui/material/CardContent'
-import List from '@mui/material/List'
+import CardHeader from '@mui/material/CardHeader'
 import Divider from '@mui/material/Divider'
+import List from '@mui/material/List'
 import { useTheme } from '@mui/material/styles'
+import React, { type FC } from 'react'
 import { makeStyles } from 'tss-react/mui'
+import { TableLine } from './table-line'
 
-interface GenericListCardProps extends PropsWithChildren {
-  title: string
+interface GenericListCardProps {
+  title: string,
+  tableLines: { label: string; value: string }[]
   ['data-testid']?: string
 }
 
@@ -57,7 +59,7 @@ const useStyles = makeStyles()((theme) => ({
 export const GenericListCard: FC<GenericListCardProps> = (props) => {
   const theme = useTheme()
   const { classes } = useStyles()
-  const { title, children } = props
+  const { title, tableLines } = props
 
   return (
     <Card
@@ -73,7 +75,9 @@ export const GenericListCard: FC<GenericListCardProps> = (props) => {
       <CardContent className={classes.cardContent}>
         <List disablePadding>
           <Divider component="li" />
-          {children}
+          {tableLines.map((item, index, array) => (
+            <TableLine key={item.label} label={item.label} value={item.value} hideDivider={index === array.length - 1} />
+          ))}
         </List>
       </CardContent>
     </Card>